Moved enums to types module
